Decouple einsum utils function to separate file
